# swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "Api Documentation",
		"version": "1.0",
		"title": "Api Documentation",
		"termsOfService": "urn:tos",
		"contact": {
			
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0"
		}
	},
	"host": "localhost:9090",
	"basePath": "/",
	"tags": [
		{
			"name": "irctc-rest-controller",
			"description": "Irctc Rest Controller"
		}
	],
	"paths": {
		"/bookticket": {
			"post": {
				"tags": [
					"irctc-rest-controller"
				],
				"summary": "bookTicket",
				"operationId": "bookTicketUsingPOST",
				"consumes": [
					"application/xml",
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"in": "body",
						"name": "request",
						"description": "request",
						"required": true,
						"schema": {
							"$ref": "#/definitions/PassengerRequest"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/ticket/{pnr}": {
			"get": {
				"tags": [
					"irctc-rest-controller"
				],
				"summary": "getTicket",
				"operationId": "getTicketUsingGET",
				"produces": [
					"application/xml",
					"application/json"
				],
				"parameters": [
					{
						"name": "pnr",
						"in": "path",
						"description": "pnr",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/TicketResponse"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		}
	},
	"definitions": {
		"PassengerRequest": {
			"type": "object",
			"properties": {
				"email": {
					"type": "string"
				},
				"from": {
					"type": "string"
				},
				"jdate": {
					"type": "string"
				},
				"name": {
					"type": "string"
				},
				"phno": {
					"type": "integer",
					"format": "int64"
				},
				"to": {
					"type": "string"
				},
				"trainNum": {
					"type": "string"
				}
			},
			"title": "PassengerRequest"
		},
		"TicketResponse": {
			"type": "object",
			"properties": {
				"cost": {
					"type": "number",
					"format": "double"
				},
				"from": {
					"type": "string"
				},
				"jdate": {
					"type": "string"
				},
				"name": {
					"type": "string"
				},
				"pnr": {
					"type": "string"
				},
				"status": {
					"type": "string"
				},
				"ticketId": {
					"type": "integer",
					"format": "int32"
				},
				"to": {
					"type": "string"
				}
			},
			"title": "TicketResponse"
		}
	}
}